Added basic workflow structure
Added facility to specify the workflow required for processing as a Protocol, and added code to run this workflow over all files in a directory.  This required significant updates to the adaptors so that file suffices could be determined.  This also adds bug-fixes for deep copying of metadata to avoid over-writing, and the ability to merge metadata rather than simply replace it.  This also normalises data to /data, assuming this will be mapped appropriately in the DevContainer specification.